Fix build errors

+++ applet/src/main/java/opencrypto/jcmathlib/ECPoint.java
@@ -938,7 +938,7 @@ public class ECPoint {
         rm.unlock(hashBuffer);
         rm.unlock(pointBuffer);
 
-        return bResult && (((short) (result & 0xffff) == (short) 0xffff));
+        return bResult && (((short) (result & (short) 0xffff) == (short) 0xffff));
     }
 
     static byte[] msg = {(byte) 0x01, (byte) 0x01, (byte) 0x02, (byte) 0x03};
